+++ packages/luciq_flutter/example/lib/main.dart
@@ -9,7 +9,6 @@ import 'package:flutter/material.dart';
 import 'package:luciq_flutter/luciq_flutter.dart';
 import 'package:luciq_flutter/src/utils/screen_loading/screen_loading_manager.dart';
 import 'package:luciq_flutter_example/src/app_routes.dart';
-import 'package:luciq_flutter_example/src/components/apm_switch.dart';
 import 'package:luciq_flutter_example/src/native/luciq_flutter_example_method_channel.dart';
 import 'package:luciq_flutter_example/src/screens/callback/callback_handler_provider.dart';
 import 'package:luciq_flutter_example/src/screens/callback/callback_page.dart';
@@ -21,12 +20,18 @@ import 'package:luciq_flutter_example/src/widget/nested_view.dart';
 import 'package:luciq_flutter_example/src/widget/section_title.dart';
 import 'package:luciq_http_client/luciq_http_client.dart';
 import 'package:provider/provider.dart';
+import 'package:luciq_flutter_example/src/utils/show_messages.dart';
 
 part 'src/components/fatal_crashes_content.dart';
 part 'src/components/flows_content.dart';
 part 'src/components/network_content.dart';
 part 'src/components/non_fatal_crashes_content.dart';
 part 'src/components/ndk_crashes_content.dart';
+part 'src/components/apm_switch.dart';
+part 'src/components/ui_traces_content.dart';
+part 'src/components/screen_render_switch.dart';
+part 'src/components/screen_render.dart';
+part 'src/components/animated_box.dart';
 
 part 'src/components/page.dart';
 part 'src/screens/apm_page.dart';
@@ -38,6 +43,7 @@ part 'src/screens/my_home_page.dart';
 part 'src/screens/screen_capture_premature_extension_page.dart';
 part 'src/screens/screen_loading_page.dart';
 part 'src/screens/session_replay_page.dart';
+part 'src/screens/screen_render_page.dart';
 
 void main() {
   runZonedGuarded(
@@ -60,9 +66,8 @@ void main() {
       );
 
       CrashReporting.setNDKEnabled(true);
-
-      CrashReporting.setNDKEnabled(true);
-
+      // APM.setScreenRenderingEnabled(true);
+      // APM.setAutoUITraceEnabled(false);
       Luciq.setWelcomeMessageMode(WelcomeMessageMode.disabled);
       FlutterError.onError = (FlutterErrorDetails details) {
         Zone.current.handleUncaughtError(details.exception, details.stack!);